pipeline: move schema to lenskit.schemas
